Fixed a bunch of assembly gen'ing bugs
Handling unsafe members with pointer types. Had to create
IsUnsafeToMock() extension methods.

Need to create the right "name" for the left-hand side of the delegate
name to include generic values (similar to creating the file name -
"RocksObjectSecurityT")

For any overridden member, if the base type or method is marked with
[ObsoleteAttribute], set a "RequiresObsoleteSuppression" flag. When
making the class, surround the class with "#pragma warning
disable/restore CS0618"

Looks like I gen'd a delete for an internal virtual method where the
method doesn't actually exist (correctly).

For some reason, not handling the constraint could be "where T : struct"